refactor: url 상수 변환 #9
