+++ src/components/ItemCard.tsx
@@ -1,4 +1,4 @@
-import { squareImgUrl } from "@/services/apiUrl";
+import { SQUARE_IMG_URL } from "@/constants/apiUrl";
 import { Item } from "@/types/Items";
 import Image from "next/image";
 
@@ -15,7 +15,7 @@ const ItemCard = ({ id, item }: ItemCardProps) => {
     >
       <div className="w-16 h-16 relative">
         <Image
-          src={`${squareImgUrl}/item/${item.image.full}`}
+          src={`${SQUARE_IMG_URL}/item/${item.image.full}`}
           alt={item.name}
           fill
           className="object-fill"
